+++ .env.sentry
@@ -1,2 +1,2 @@
-SENTRY_RELEASE=release-00012
-VITE_SENTRY_RELEASE=release-00012
+SENTRY_RELEASE=release-00014
+VITE_SENTRY_RELEASE=release-00014
