+++ src/components/PanelDesigner/types.ts
@@ -1,0 +1,3 @@
+import type { useI18n } from '@i18n/I18nContext';
+
+export type ReturnTypeUseI18n = ReturnType<typeof useI18n>;
